# academe/server/openapi.yaml
openapi: 3.1.0
info:
  title: academe API
  version: "1"
  description: |
    Every error is {"error": {"code", "message", "requestId"}}. The app
    switches on `code`; `message` is for people.
    Access tokens last 15 minutes. When one is rejected with `invalid_token`,
    call /auth/refresh once; if that is rejected too, log the student out.
servers:
  - url: http://localhost:8080
paths:
...
  /auth/sign-up:
    post:
      summary: Create an account with email and password, and log it in. 10 an hour per IP address.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignUp" }
      responses:
        "201":
          description: Created and logged in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "409":
          description: "`email_taken`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_firstName`, `invalid_lastName`, `invalid_email` or `invalid_password`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { $ref: "#/components/responses/Throttled" }
  /auth/log-in:
    post:
      summary: Log in with email and password. 10 per 15 minutes per email and IP address, 50 an hour per IP address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              additionalProperties: false
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: "`wrong_credentials` — never says whether the email exists."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { $ref: "#/components/responses/Throttled" }
  /auth/refresh:
    post:
      summary: Trade a refresh token for new tokens. The old one stops working.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshToken" }
      responses:
        "200":
          description: New tokens.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tokens" }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: "`invalid_token` — expired, used or logged out. Log the student out."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/log-out:
    post:
      summary: End the session a refresh token belongs to. Unknown tokens are ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshToken" }
      responses:
        "204": { description: Logged out. }
        "400": { $ref: "#/components/responses/Error" }
  /auth/google:
    post:
      summary: Sign in with a Google ID token. Creates the account on first use; links an existing email account (its password and other sessions are dropped). 30 an hour per IP address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              additionalProperties: false
              properties:
                idToken: { type: string }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Session"
                  - type: object
                    required: [created]
                    properties:
                      created: { type: boolean, description: True when this call created the account. }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: "`invalid_google_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: "`email_taken`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "503":
          description: "`google_unavailable` — the server has no Google client IDs."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { $ref: "#/components/responses/Throttled" }
  /auth/password-reset:
    post:
      summary: Email a 6-digit reset code and a Reset password link. Google-only accounts get an email saying to use Continue with Google. Same answer whether or not the email has an account.
      description: At most 3 requests per email and 10 per IP address an hour. The code and the link last 15 minutes and share the code's 5 tries; asking again replaces both. The link is https://academe.cc/reset-password?c=<linkToken>, where linkToken is 128 random bits in unpadded base64url (stored hashed, single use).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              additionalProperties: false
              properties:
                email: { type: string }
      responses:
        "202":
          description: Accepted. Always the same body.
          content:
            application/json:
              schema:
                type: object
                required: [expiresIn]
                properties:
                  expiresIn: { type: integer, description: Code lifetime in seconds. }
        "400": { $ref: "#/components/responses/Error" }
        "429":
          description: "`too_many_requests`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/password-reset/verify:
    post:
      summary: Check an emailed code and trade it for a single-use reset token. Five tries per code.
      description: At most 30 tries per IP address an hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              additionalProperties: false
              properties:
                email: { type: string }
                code: { type: string, description: Six digits; spaces are ignored. }
      responses:
        "200":
          description: The code was right.
          content:
            application/json:
              schema:
                type: object
                required: [resetToken, expiresIn]
                properties:
                  resetToken: { type: string }
                  expiresIn: { type: integer, description: Reset token lifetime in seconds. }
        "400": { $ref: "#/components/responses/Error" }
        "410":
          description: "`code_expired` — expired, used, replaced, or never sent. Ask for a new code."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`wrong_code`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: "`too_many_attempts` — the code is burned, ask for a new one; or `too_many_requests`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/password-reset/complete:
    post:
      summary: Set a new password with a reset token. Ends every other session, makes every earlier access token answer 401, and logs the student in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, password]
              additionalProperties: false
              properties:
                resetToken: { type: string }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "200":
          description: Password changed and logged in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "410":
          description: "`reset_expired` — the token expired or was used. Start again."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_password`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/password-reset/link:
    post:
      summary: Trade the link token from a reset email (the c parameter of https://academe.cc/reset-password or academe://reset) for a single-use reset token, then call /auth/password-reset/complete.
      description: Single use; the emailed code stops working too. Counts as one of the code's 5 tries and against the 30 tries per IP address an hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [linkToken]
              additionalProperties: false
              properties:
                linkToken: { type: string, maxLength: 64 }
      responses:
        "200":
          description: A reset token for /auth/password-reset/complete.
          content:
            application/json:
              schema:
                type: object
                required: [resetToken, expiresIn]
                properties:
                  resetToken: { type: string }
                  expiresIn: { type: integer, description: Token lifetime in seconds. }
        "400": { $ref: "#/components/responses/Error" }
        "410":
          description: "`reset_expired` — the link is unknown, expired, used, or replaced by a newer email."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: "`too_many_attempts` (the code's tries are used up) or `too_many_requests`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /me:
    patch:
      summary: Change the logged-in student's name.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName]
              additionalProperties: false
              properties:
                firstName: { type: string, maxLength: 100 }
                lastName: { type: string, maxLength: 100 }
      responses:
        "200":
          description: The updated account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: "`invalid_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_firstName` or `invalid_lastName`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    get:
      summary: The logged-in student's account.
      security: [{ bearer: [] }]
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "401":
          description: "`invalid_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    delete:
      summary: Schedule the student's account for deletion. Sessions end now; everything (profile, XP, chats) is erased for good after 30 days. Logging in before then cancels it. The body is optional.
      security: [{ bearer: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason: { type: string, maxLength: 500, description: Why they're leaving; optional. }
      responses:
        "202":
          description: Scheduled.
          content:
            application/json:
              schema:
                type: object
                required: [deletesAt]
                properties:
                  deletesAt: { type: string, format: date-time }
        "401":
          description: "`invalid_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_reason`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /me/password:
    post:
      summary: Change the password, or set the first one on a Google-only account (then email log-in works too). Every other session and access token is revoked; the response carries fresh tokens for this device.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              additionalProperties: false
              properties:
                currentPassword: { type: string, description: Required when the account has a password; ignored when it has none. }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "200":
          description: Changed. Replace the stored tokens with these.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: "`invalid_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "403":
          description: "`wrong_password` — the current password is wrong."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_password` — the new password breaks the sign-up rules."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: "`too_many_requests` — five current-password checks an hour per account."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /me/google:
    post:
      summary: Link a Google account to the logged-in account. The password and sessions are kept.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              additionalProperties: false
              properties:
                idToken: { type: string, description: A Google ID token, as for /auth/google. }
      responses:
        "200":
          description: Linked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/Error" }
        "401":
          description: "`invalid_token` or `invalid_google_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: "`google_taken` — that Google account is linked to another account."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "503":
          description: "`google_unavailable` — the server has no Google client IDs."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    delete:
      summary: Unlink Google. Only allowed when the account has a password.
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Unlinked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "401":
          description: "`invalid_token`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: "`password_required` — set a password first."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: The access token from a Session. Any bearer route answers 401 `invalid_token` when the token has expired (15 minutes) or was issued before a password reset, a Google link or an account deletion request; refresh or log in again. Other server instances may accept a revoked token for up to 30 seconds.
  responses:
...
    Error:
      description: "`invalid_json`, `body_too_large` or `internal`"
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Throttled:
      description: "`too_many_requests` — this IP address (IPv6 grouped by /64), or this email from this IP address, tried too often."
      headers:
        Retry-After: { description: Seconds until the next try is allowed., schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    SignUp:
      type: object
      required: [firstName, lastName, email, password]
      additionalProperties: false
      properties:
        firstName: { type: string, maxLength: 100, description: Trimmed; must not be blank. }
        lastName: { type: string, maxLength: 100, description: Trimmed; must not be blank. }
        email: { type: string, maxLength: 254, description: Trimmed and lower-cased. }
        password: { type: string, minLength: 8, maxLength: 128 }
    RefreshToken:
      type: object
      required: [refreshToken]
      additionalProperties: false
      properties:
        refreshToken: { type: string }
    Account:
      type: object
      required: [id, firstName, lastName, email]
      properties:
        id: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        hasPassword: { type: boolean, description: "False for a Google-only account: Account settings offers \"Set a password\" instead of \"Change password\", and Google can't be unlinked." }
        googleEmail: { type: string, description: The linked Google account's email. Absent when Google isn't linked. }
    Tokens:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string, description: "Send as `Authorization: Bearer …`." }
        refreshToken: { type: string, description: Keep in secure storage; lasts 30 days. }
        expiresIn: { type: integer, description: Access token lifetime in seconds. }
    Session:
      type: object
      required: [account, tokens]
      properties:
        account: { $ref: "#/components/schemas/Account" }
        tokens: { $ref: "#/components/schemas/Tokens" }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, description: "Only on some errors, e.g. limit_reached and pro_only: {feature, limit, resetsAt}." }